refactorizado deleteProduct
